+++ src/index/datacore.ts
@@ -65,11 +65,31 @@ export class Datacore extends Component {
 
     /** Initialize datacore by scanning persisted caches and all available files, and queueing parses as needed. */
     initialize() {
-        // The metadata cache is updated on initial file index and file loads.
+        // Metadata cache handles markdown file updates.
         this.registerEvent(this.metadataCache.on("resolve", (file) => this.reload(file)));
 
         // Renames do not set off the metadata cache; catch these explicitly.
         this.registerEvent(this.vault.on("rename", this.rename, this));
+
+        // Handle generic file creates and updates; resolve generally only applies to markdown files
+        // but we do keep basic metadata about all files.
+        this.registerEvent(this.vault.on("create", (file) => {
+            if (!(file instanceof TFile)) return;
+
+            // Handled by the metadata cache.
+            if (file.extension === "md" || file.extension === "markdown") return;
+
+            this.reload(file);
+        }));
+
+        this.registerEvent(this.vault.on("modify", (file) => {
+            if (!(file instanceof TFile)) return;
+
+            // Handled by the metadata cache.
+            if (file.extension === "md" || file.extension === "markdown") return;
+
+            this.reload(file);
+        }));
 
         // File creation does cause a metadata change, but deletes do not. Clear the caches for this.
         this.registerEvent(
@@ -100,6 +120,7 @@ export class Datacore extends Component {
 
             this.datastore.touch();
             this.trigger("update", this.revision);
+            this.trigger("initialized");
 
             // Clean up any documents which no longer exist in the vault.
             // TODO: I think this may race with other concurrent operations, so
@@ -123,6 +144,9 @@ export class Datacore extends Component {
         // (for sections, tasks, etc to refer to their parent file) and it requires some finesse to fix.
         this.datastore.delete(oldPath);
         this.reload(file);
+
+        // TODO: For correctness, probably have to either fix links in all linked files OR
+        // just stop normalizing links in the store.
     }
 
     /**
@@ -194,6 +218,7 @@ export class Datacore extends Component {
 
     /** Called whenever the index updates to a new revision. This is the broadest possible datacore event. */
     public on(evt: "update", callback: (revision: number) => any, context?: any): EventRef;
+    public on(evt: "initialized", callback: () => any, context?: any): EventRef;
 
     on(evt: string, callback: (...data: any) => any, context?: any): EventRef {
         return this.events.on(evt, callback, context);
@@ -211,6 +236,8 @@ export class Datacore extends Component {
 
     /** Trigger an update event. */
     private trigger(evt: "update", revision: number): void;
+    /** Trigger an initialization event. */
+    private trigger(evt: "initialized"): void;
 
     /** Trigger an event. */
     private trigger(evt: string, ...args: any[]): void {
@@ -233,6 +260,8 @@ export class DatacoreInitializer extends Component {
     /** Deferred promise which resolves when importing is done. */
     done: Deferred<InitializationStats>;
 
+    /** The total number of target files to import. */
+    targetTotal: number;
     /** The time that init started in milliseconds. */
     start: number;
     /** Total number of files to import. */
@@ -251,6 +280,7 @@ export class DatacoreInitializer extends Component {
 
         this.active = false;
         this.queue = this.core.vault.getFiles();
+        this.targetTotal = this.queue.length;
         this.files = this.queue.length;
         this.start = Date.now();
         this.current = [];
@@ -333,14 +363,11 @@ export class DatacoreInitializer extends Component {
                     this.core.storeMarkdown(data);
                     return { status: "cached" };
                 }
-
-                // Does not match an existing import type, just reload normally.
-                await this.core.reload(file);
-                return { status: "imported" };
-            } else {
-                await this.core.reload(file);
-                return { status: "imported" };
             }
+
+            // Does not match an existing import type, just reload normally.
+            await this.core.reload(file);
+            return { status: "imported" };
         } catch (ex) {
             console.log("Datacore: Failed to import file: ", ex);
             return { status: "skipped" };